Split pop controller endpoints by domain
